improve props
