+++ app/components/common/BaseRawDetails.tsx
@@ -5,8 +5,12 @@ import React from 'react';
 import { AddressFromLookupTableWithContext, AddressWithContext } from '../inspector/AddressWithContext';
 import { fillAddressTableLookupsAccounts, findLookupAddress } from '../inspector/utils';
 import { HexData } from './HexData';
-import { AddressTableLookupAddress } from './inspector/AddressTableLookupAddress';
 
+/**
+ *  Component that displays accounts from any Instruction.
+ *
+ *  VersionedMessage is optional as it will be present at inspector page only.
+ */
 export function BaseRawDetails({
     ix,
     message,
