+++ app/components/inspector/associated-token/RecoverNestedDetailsCard.tsx
@@ -1,13 +1,7 @@
 import { Address } from '@components/common/Address';
 import { AddressTableLookupAddress } from '@components/common/inspector/AddressTableLookupAddress';
 import { InspectorInstructionCard } from '@components/common/InspectorInstructionCard';
-import {
-    MessageCompiledInstruction,
-    ParsedInstruction,
-    SignatureResult,
-    TransactionInstruction,
-    VersionedMessage,
-} from '@solana/web3.js';
+import { MessageCompiledInstruction, ParsedInstruction, SignatureResult, VersionedMessage } from '@solana/web3.js';
 import React from 'react';
 
 export function RecoverNestedDetailsCard(props: {
@@ -15,8 +9,8 @@ export function RecoverNestedDetailsCard(props: {
     index: number;
     innerCards?: JSX.Element[];
     ix: ParsedInstruction;
-    message?: VersionedMessage;
-    raw: TransactionInstruction | MessageCompiledInstruction;
+    message: VersionedMessage;
+    raw: MessageCompiledInstruction;
     result: SignatureResult;
     InstructionCardComponent?: React.FC<Parameters<typeof InspectorInstructionCard>[0]>;
 }) {
@@ -30,11 +24,6 @@ export function RecoverNestedDetailsCard(props: {
         childIndex,
         InstructionCardComponent = InspectorInstructionCard,
     } = props;
-
-    let accountKeys;
-    if ('accountKeyIndexes' in raw) {
-        accountKeys = raw.accountKeyIndexes;
-    }
 
     return (
         <InstructionCardComponent
@@ -56,57 +45,43 @@ export function RecoverNestedDetailsCard(props: {
             <tr>
                 <td>Destination</td>
                 <td className="text-lg-end">
-                    {message && accountKeys ? (
-                        <AddressTableLookupAddress accountIndex={accountKeys[2]} message={message} hideInfo />
-                    ) : null}
+                    <AddressTableLookupAddress accountIndex={raw.accountKeyIndexes[2]} message={message} hideInfo />
                 </td>
             </tr>
             <tr>
                 <td>Nested Mint</td>
                 <td className="text-lg-end">
-                    {message && accountKeys ? (
-                        <AddressTableLookupAddress accountIndex={accountKeys[1]} message={message} hideInfo />
-                    ) : null}
+                    <AddressTableLookupAddress accountIndex={raw.accountKeyIndexes[1]} message={message} hideInfo />
                 </td>
             </tr>
             <tr>
                 <td>Nested Owner</td>
                 <td className="text-lg-end">
-                    {message && accountKeys ? (
-                        <AddressTableLookupAddress accountIndex={accountKeys[3]} message={message} hideInfo />
-                    ) : null}
+                    <AddressTableLookupAddress accountIndex={raw.accountKeyIndexes[3]} message={message} hideInfo />
                 </td>
             </tr>
             <tr>
                 <td>Nested Source</td>
                 <td className="text-lg-end">
-                    {message && accountKeys ? (
-                        <AddressTableLookupAddress accountIndex={accountKeys[0]} message={message} hideInfo />
-                    ) : null}
+                    <AddressTableLookupAddress accountIndex={raw.accountKeyIndexes[0]} message={message} hideInfo />
                 </td>
             </tr>
             <tr>
                 <td>Owner Mint</td>
                 <td className="text-lg-end">
-                    {message && accountKeys ? (
-                        <AddressTableLookupAddress accountIndex={accountKeys[4]} message={message} hideInfo />
-                    ) : null}
+                    <AddressTableLookupAddress accountIndex={raw.accountKeyIndexes[4]} message={message} hideInfo />
                 </td>
             </tr>
             <tr>
                 <td>Owner</td>
                 <td className="text-lg-end">
-                    {message && accountKeys ? (
-                        <AddressTableLookupAddress accountIndex={accountKeys[5]} message={message} hideInfo />
-                    ) : null}
+                    <AddressTableLookupAddress accountIndex={raw.accountKeyIndexes[5]} message={message} hideInfo />
                 </td>
             </tr>
             <tr>
                 <td>Token Program</td>
                 <td className="text-lg-end">
-                    {message && accountKeys ? (
-                        <AddressTableLookupAddress accountIndex={accountKeys[6]} message={message} hideInfo />
-                    ) : null}
+                    <AddressTableLookupAddress accountIndex={raw.accountKeyIndexes[6]} message={message} hideInfo />
                 </td>
             </tr>
         </InstructionCardComponent>
